+++ typespec-extension/src/type-utils.ts
@@ -25,6 +25,7 @@ import { SchemaContext } from "@autorest/codemodel";
 import { DurationSchema } from "./common/schemas/time.js";
 import { getNamespace, pascalCase } from "./utils.js";
 import { getUnionAsEnum } from "@azure-tools/typespec-azure-core";
+import { Version } from "@typespec/versioning";
 
 /** Acts as a cache for processing inputs.
  *
@@ -49,6 +50,10 @@ export class ProcessingCache<In, Out> {
     }
     return undefined;
   }
+}
+
+export function isStable(version: Version): boolean {
+  return !version.value.toLowerCase().includes("preview");
 }
 
 /** adds only if the item is not in the collection already